also restructured the communities api
