+++ server/api/communities/communities.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
-const { Community, User, CommunityTask, Task, TaskItem } = require('../db/models')
-const sendEmail = require('../mailer')
+const { Community, User, TaskItem } = require('../../db/models')
+const sendEmail = require('../../mailer')
 module.exports = router
 
 //mailer takes --> email, user, community, status
